feat: SaveDataManagerをworld_1ディレクトリ丸ごとのバックアップへ移行
旧実装はSaves直下のファイルだけを見ており、ワールドディレクトリ形式へ
移行済みの現行セーブ(world_1/)を1バイトもバックアップしていなかった。

- Editor/SaveData/ へ移設し、UI(SaveDataManager) / ダイアログ / IO(WorldBackupService)へ分割
- Backup_<日時>_<説明>/world_1/ の構造で除外なしに再帰コピー
- 削除はworld_1ディレクトリごと。旧形式のsave_N.jsonは対象外
- 既定ワールド名の正をGameSystemPaths.DefaultWorldNameへ集約
