+++ .decisions/2026-08-18-セーブバックアップはworld_1固定でディレクトリ丸ごと.md
@@ -1,0 +1,8 @@
+決定: SaveDataManagerのバックアップ対象をworld_1固定とし、`Backup_<日時>_<説明>/world_1/`の構造で除外なしに丸ごとコピーする。Backup&Delete/Deleteの削除粒度はworld_1ディレクトリごと削除。Saves直下に残る旧形式ファイル(save_1.json等)はバックアップも削除もせず完全に無視する。
+棄却案:
+- Saves直下のworld_*を全部バックアップ（world_generatedが単体330MB・terrain 300MBで1ボタンに重すぎる。実プレイのワールドはworld_1のみ）
+- ウィンドウでワールドを選択させる（UIを足す割に対象がworld_1で足りる）
+- terrain/visual/cacheを除外して軽量化（seedから再生成可能だが復元手順が増える）
+- 削除はsave.jsonだけにして地形を残す（ワールド単位で作り直す運用に合わない）
+理由: 旧実装はSaves直下のファイルしか見ておらず現行セーブを1バイトも保存していなかった。対象をworld_1固定・除外なしに絞ることで、確実に復元できるバックアップが最小のUIで得られる。
+リンク: 実装は`moorestech_server/Assets/Scripts/Editor/SaveData/`、既定ワールド名の正は`GameSystemPaths.DefaultWorldName`
